chore: remove StartLiveness, refactor
