--- packages/react-liveness/src/components/FaceLivenessDetector/StartLiveness/index.tsx
@@ -1,1 +1,0 @@
-export * from './StartLiveness';
