+++ packages/react-liveness/src/components/FaceLivenessDetector/utils/helpers.ts
